hopefully sync the branch

+++ core/tests/loadnext/src/executor.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use anyhow::anyhow;
 use futures::{channel::mpsc, future, SinkExt};
-use zksync_eth_client::Options;
+use zksync_eth_client::{EthInterface, Options};
 use zksync_eth_signer::PrivateKeySigner;
 use zksync_system_constants::MAX_L1_TRANSACTION_GAS_LIMIT;
 use zksync_types::{
@@ -117,7 +117,7 @@ impl Executor {
         );
         LOADTEST_METRICS
             .master_account_balance
-            .set(eth_balance.as_u128() as u64);
+            .set(eth_balance.as_u128() as f64);
 
         Ok(())
     }
